+++ src/operandomerge/__init__.py
@@ -23,4 +23,4 @@ __all__ = [
     "MergeService",
     "TimeRepresentation",
 ]
-__version__ = "0.1.2"
+__version__ = "0.1.3"